micro-backend/requirements-separation.txt
torch
torchaudio
asteroid-filterbanks
huggingface-hub
PyYAML
SoundFile

# Windows note: install asteroid separately without optional evaluation deps:
# .\.venv-win\python.exe -m pip install --no-deps asteroid
